backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
opencv-python-headless==4.11.0.86
pytesseract==0.3.13
pandas==2.2.3
openpyxl==3.1.5
numpy==2.2.5
Pillow==11.2.1
google-cloud-vision==3.10.2
pydantic-settings==2.9.1
